modularized bot
